fix(backend): prevent permission elevation and unauthorized access in share services
- Fix permission elevation vulnerability in base_service.py: only use requested_permission_level when require_approval is True
- Fix redundant exception logging in share_webhook.py: remove formatted exception from logger.exception calls
- Fix unauthorized access in task_share_service.py: return None instead of task when user lacks access
- Add type annotation to team_share_service.py __init__ method
- Fix unauthorized access in team_share_service.py: return None instead of team when user lacks access

fix(frontend): improve code quality and security

- Rename AddUserDialog.tsx to add-user-dialog.tsx (kebab-case)
- Fix React.FormEvent type import: import FormEvent directly from React
- Extract form UI into AddUserForm subcomponent to keep file under 50 lines
- Fix useEffect closure dependency in PermissionManagementTab: use functional update pattern
- Replace unsafe wildcard '*' with explicit 'localhost:3000' in next.config.js
